requirements.txt
dhanhq>=2.1.0
pandas>=1.5.0
numpy>=1.24.0
python-dotenv>=0.21.0
requests>=2.28.0
yfinance>=0.2.18
schedule>=1.2.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0